Shared types directory
commit d386f3d17fbae9059fbf3e9add6af91eefeb011e
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Fri Apr 4 16:51:15 2025 +0300

    shared types for services

commit 63a384a7f31c948bf67511247570f8f82f6d38ba
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Fri Apr 4 16:04:21 2025 +0300

    shared types folder

commit 33c8ed69e11eb6a3f057a22d4d02ca335206d8c9
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Fri Apr 4 15:21:12 2025 +0300

    wip tsconfig
