+++ backend/src/models/attendanceModel.ts
@@ -1,5 +1,5 @@
+import { Attendance, AttendanceModelForCreation } from "@shared/index";
 import { sql } from "bun";
-import { t } from "elysia";
 
 /**
  * Event Attendance Data Transfer Object
@@ -15,31 +15,3 @@ export const AttendanceDTO = {
     return newAttendance;
   },
 };
-
-export const AttendanceModelForCreation = t.Object({
-  event_id: t.Integer(),
-  user_id: t.Integer(),
-  status: t
-    .Enum({
-      accepted: "accepted",
-      declined: "declined",
-      tentative: "tentative",
-      "needs-action": "needs-action",
-    })
-    .Default("needs-action"),
-});
-export type AttendanceModelForCreation =
-  typeof AttendanceModelForCreation.static;
-
-export const AttendanceModel = t.Object({
-  id: t.Integer(),
-  event_id: t.Integer(),
-  user_id: t.Integer(),
-  status: t.Enum({
-    accepted: "accepted",
-    declined: "declined",
-    tentative: "tentative",
-    "needs-action": "needs-action",
-  }),
-});
-export type Attendance = typeof AttendanceModel.static;
