+++ shared/src/types.ts
@@ -1,0 +1,1 @@
+export const placeholder: number = 1;
